Moved the entryWindow to a new class file. Also changed to an ArrayList
