+++ task_list/src/app/taskEntry.java
@@ -75,4 +75,14 @@ public class taskEntry {
 		}
 		return "Not Started";
 	}
+	
+	public void decrementPriority()
+	{
+		priority--;
+	}
+	
+	public void incrementPriority()
+	{
+		priority++;
+	}
 }
